feat: finishing homework 1

+++ Doctest_tests/ArrayTest.cpp
@@ -15,13 +15,13 @@ TEST_CASE("Array::Array()") {
     array = Array(2);
     CHECK(array.size == 2);
 
-    array = Array(std::vector<double>{1, 2, 3});
+    array = Array(std::vector<long double>{1, 2, 3});
     CHECK(array.size == 3);
-    CHECK(array.array == std::vector<double>{1, 2, 3});
+    CHECK(array.array == std::vector<long double>{1, 2, 3});
 
     array = Array("[1 2 3]");
     CHECK(array.size == 3);
-    CHECK(array.array == std::vector<double>{1, 2, 3});
+    CHECK(array.array == std::vector<long double>{1, 2, 3});
 }
 
 TEST_CASE("Array::print()") {
@@ -34,14 +34,14 @@ TEST_CASE("Array::operator+()") {
     array1 = Array("[1 2 3]");
     array2 = Array("[4 5 6]");
     array3 = array1 + array2;
-    CHECK(array3.array == std::vector<double>{5, 7, 9});
+    CHECK(array3.array == std::vector<long double>{5, 7, 9});
     array3 = array2 - array1;
-    CHECK(array3.array == std::vector<double>{3, 3, 3});
+    CHECK(array3.array == std::vector<long double>{3, 3, 3});
     array3 = array1 * 2;
-    CHECK(array3.array == std::vector<double>{2, 4, 6});
+    CHECK(array3.array == std::vector<long double>{2, 4, 6});
     array3 = array1 / 2;
-    CHECK(array3.array == std::vector<double>{0.5, 1, 1.5});
-    double dot_product = array1 * array2;
+    CHECK(array3.array == std::vector<long double>{0.5, 1, 1.5});
+    long double dot_product = array1 * array2;
     CHECK(dot_product == 32);
 
     array3 = Array("[1 2 3 4 5 6 7 8 9 10]");
