+++ Doctest_tests/MatrixTest.cpp
@@ -10,26 +10,26 @@ TEST_CASE("Matrix::Matrix()") {
     Matrix matrix;
     CHECK(matrix.rows == 0);
     CHECK(matrix.cols == 0);
-    CHECK(matrix.matrix == std::vector<std::vector<double>>(0, std::vector<double>(0, 0)));
+    CHECK(matrix.matrix == std::vector<std::vector<long double>>(0, std::vector<long double>(0, 0)));
 
     matrix = Matrix(2, 3);
     CHECK(matrix.rows == 2);
     CHECK(matrix.cols == 3);
-    CHECK(matrix.matrix == std::vector<std::vector<double>>(2, std::vector<double>(3, 0)));
+    CHECK(matrix.matrix == std::vector<std::vector<long double>>(2, std::vector<long double>(3, 0)));
 
-    matrix = Matrix(std::vector<std::vector<double>>({{1, 2, 3},
-                                                      {4, 5, 6}}));
+    matrix = Matrix(std::vector<std::vector<long double>>({{1, 2, 3},
+                                                           {4, 5, 6}}));
     CHECK(matrix.rows == 2);
     CHECK(matrix.cols == 3);
-    CHECK(matrix.matrix == std::vector<std::vector<double>>({{1, 2, 3},
-                                                             {4, 5, 6}}));
+    CHECK(matrix.matrix == std::vector<std::vector<long double>>({{1, 2, 3},
+                                                                  {4, 5, 6}}));
 
     matrix = Matrix("[1 2 3; 4 5 6; 7 8 9]");
     CHECK(matrix.rows == 3);
     CHECK(matrix.cols == 3);
-    CHECK(matrix.matrix == std::vector<std::vector<double>>({{1, 2, 3},
-                                                             {4, 5, 6},
-                                                             {7, 8, 9}}));
+    CHECK(matrix.matrix == std::vector<std::vector<long double>>({{1, 2, 3},
+                                                                  {4, 5, 6},
+                                                                  {7, 8, 9}}));
 }
 
 TEST_CASE("Matrix::print") {
@@ -42,29 +42,29 @@ TEST_CASE("Matrix::operators") {
     m1 = Matrix("[1 2 3; 4 5 6; 7 8 9]");
     m2 = Matrix("[1 0 0; 0 1 0; 0 0 1]"); // Identity matrix
     m3 = m1 + m2;
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{2, 2, 3},
-                                                         {4, 6, 6},
-                                                         {7, 8, 10}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{2, 2, 3},
+                                                              {4, 6, 6},
+                                                              {7, 8, 10}}));
     m3 = m1 - m2;
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{0, 2, 3},
-                                                         {4, 4, 6},
-                                                         {7, 8, 8}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{0, 2, 3},
+                                                              {4, 4, 6},
+                                                              {7, 8, 8}}));
     m3 = m1 * m1;
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{30,  36,  42},
-                                                         {66,  81,  96},
-                                                         {102, 126, 150}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{30,  36,  42},
+                                                              {66,  81,  96},
+                                                              {102, 126, 150}}));
     m3 = m1.transpose();
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{1, 4, 7},
-                                                         {2, 5, 8},
-                                                         {3, 6, 9}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{1, 4, 7},
+                                                              {2, 5, 8},
+                                                              {3, 6, 9}}));
     m3 = m1 * 2;
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{2,  4,  6},
-                                                         {8,  10, 12},
-                                                         {14, 16, 18}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{2,  4,  6},
+                                                              {8,  10, 12},
+                                                              {14, 16, 18}}));
     m3 = m1 / 2;
-    CHECK(m3.matrix == std::vector<std::vector<double>>({{0.5, 1,   1.5},
-                                                         {2,   2.5, 3},
-                                                         {3.5, 4,   4.5}}));
+    CHECK(m3.matrix == std::vector<std::vector<long double>>({{0.5, 1,   1.5},
+                                                              {2,   2.5, 3},
+                                                              {3.5, 4,   4.5}}));
 
     m2 = Matrix("[1 2 3; 4 5 6]");
     CHECK_THROWS(m3 = m1 + m2);
@@ -79,9 +79,9 @@ TEST_CASE("Matrix * Array") {
     m1 = Matrix("[1 2 3; 4 5 6; 7 8 9]");
     a1 = Array("[1 2 3]");
     a2 = m1 * a1;
-    CHECK(a2.array == std::vector<double>{14, 32, 50});
+    CHECK(a2.array == std::vector<long double>{14, 32, 50});
     a2 = a1 * m1;
-    CHECK(a2.array == std::vector<double>{30, 36, 42});
+    CHECK(a2.array == std::vector<long double>{30, 36, 42});
     a1 = Array("[1 2 3 4 5 6 7 8 9 10]");
     CHECK_THROWS(a2 = m1 * a1);
     CHECK_THROWS(a2 = a1 * m1);
@@ -89,7 +89,7 @@ TEST_CASE("Matrix * Array") {
     a1 = Array("[1 2 3]");
     a2 = Array("[1 2 3]");
     m1 = Matrix::product(a1, a2);
-    CHECK(m1.matrix == std::vector<std::vector<double>>({{1, 2, 3},
-                                                         {2, 4, 6},
-                                                         {3, 6, 9}}));
+    CHECK(m1.matrix == std::vector<std::vector<long double>>({{1, 2, 3},
+                                                              {2, 4, 6},
+                                                              {3, 6, 9}}));
 }
